Create a seperate file to hold shared definitions. (#147)
